Repo Fixed
Github is weird. When I try to make a new repo in a location, it creates a folder inside that location, then puts the git info in there. I have to manually put all the git info in the new folder into the parent folder (where the project files are), and then tell git the path again. Frustrating.